+++ src/Notify/assets/src/controller.ts
@@ -49,7 +49,9 @@ export default class extends Controller {
         }
 
         this.eventSources.forEach((eventSource) => {
-            const listener = (event: MessageEvent) => this._notify(JSON.parse(event.data).summary);
+            const listener = (event: MessageEvent) =>
+                this._notify(JSON.parse(event.data).summary, JSON.parse(event.data).content);
+
             eventSource.addEventListener('message', listener);
             this.listeners.set(eventSource, listener);
         });
@@ -70,11 +72,11 @@ export default class extends Controller {
         this.eventSources = [];
     }
 
-    _notify(content: string | undefined) {
-        if (!content) return;
+    _notify(title: string | undefined, options: NotificationOptions | undefined) {
+        if (!title) return;
 
         if ('granted' === Notification.permission) {
-            new Notification(content);
+            new Notification(title, options);
 
             return;
         }
@@ -82,7 +84,7 @@ export default class extends Controller {
         if ('denied' !== Notification.permission) {
             Notification.requestPermission().then((permission) => {
                 if ('granted' === permission) {
-                    new Notification(content);
+                    new Notification(title, options);
                 }
             });
         }